Cleanup edgeless_function
