+++ edgeless_node/Cargo.toml
@@ -37,7 +37,6 @@ serde_json = "1.0"
 clap = { version = "4.3", features=["derive"] }
 toml = "0.7"
 edgeless_api = {path = "../edgeless_api", features = ["grpc_impl"]}
-edgeless_function = { path = "../edgeless_function" }
 edgeless_telemetry = { path = "../edgeless_telemetry" }
 edgeless_dataplane = { path = "../edgeless_dataplane" }
 edgeless_http = {path = "../edgeless_http", features = ["hyper"]}
